requirements.txt
llama_index
pymongo
openai
langchain
python_dotenv
motor
langchain-community